[cc1] Remove old time bits from types
This short type there wass from ancient times when the types
were used in a different way.

+++ cc1/decl.c
@@ -761,7 +761,7 @@ identifier(struct decl *dcl)
 	if (sym == NULL) {
 		sym = redcl(dcl->sym, tp, dcl->pars, sclass);
 	} else {
-		short flags = sym->flags;
+		int flags = sym->flags;
 
 		sym->type = tp;
 		sym->u.pars = dcl->pars;
